backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
qrcode[pil]>=7.4.2
matplotlib>=3.7.0
scipy>=1.11.0
reportlab>=4.0.0
ultralytics>=8.0.0
pandas>=2.0.0
huggingface_hub
